// model/createGeneralLedgerAccountsRequest.ts
/**
 * ManageSpace Accounting API
 * ManageSpace Accounting API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CreateGeneralLedgerAccountRequest } from './createGeneralLedgerAccountRequest';

export class CreateGeneralLedgerAccountsRequest {
    /**
    * General ledger account
    */
    'generalLedgerAccounts': Array<CreateGeneralLedgerAccountRequest>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "generalLedgerAccounts",
            "baseName": "generalLedgerAccounts",
            "type": "Array<CreateGeneralLedgerAccountRequest>"
        }    ];

    static getAttributeTypeMap() {
        return CreateGeneralLedgerAccountsRequest.attributeTypeMap;
    }
}