// model/createGeneralLedgerEntryRequest.ts
/**
 * ManageSpace Accounting API
 * ManageSpace Accounting API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CreateGeneralLedgerLineItemRequest } from './createGeneralLedgerLineItemRequest';
import { GeneralLedgerEntryType } from './generalLedgerEntryType';

export class CreateGeneralLedgerEntryRequest {
    /**
    * ID of the Ledger
    */
    'ledgerId': string;
    /**
    * ID of the Location
    */
    'locationId': string;
    /**
    * description of the entry
    */
    'description'?: string;
    /**
    * tag of the entry
    */
    'tag': string;
    'generalLedgerEntryType': GeneralLedgerEntryType;
    /**
    * The date the change was made.
    */
    'entryDate': Date;
    /**
    * The line items of the entry
    */
    'lineItems': Array<CreateGeneralLedgerLineItemRequest>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "ledgerId",
            "baseName": "ledgerId",
            "type": "string"
        },
        {
            "name": "locationId",
            "baseName": "locationId",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "tag",
            "baseName": "tag",
            "type": "string"
        },
        {
            "name": "generalLedgerEntryType",
            "baseName": "generalLedgerEntryType",
            "type": "GeneralLedgerEntryType"
        },
        {
            "name": "entryDate",
            "baseName": "entryDate",
            "type": "Date"
        },
        {
            "name": "lineItems",
            "baseName": "lineItems",
            "type": "Array<CreateGeneralLedgerLineItemRequest>"
        }    ];

    static getAttributeTypeMap() {
        return CreateGeneralLedgerEntryRequest.attributeTypeMap;
    }
}

export namespace CreateGeneralLedgerEntryRequest {
}